+++ Modules/ComputerManagementDsc/Examples/Resources/WindowsEventlog/1-WindowsEventlog_SetWindowsEventlogSize_Config.ps1
@@ -1,0 +1,39 @@
+<#PSScriptInfo
+.VERSION 1.0.0
+.GUID f8fb71fd-9f4a-4ae5-93b8-53362752e37d
+.AUTHOR Microsoft Corporation
+.COMPANYNAME Microsoft Corporation
+.COPYRIGHT (c) Microsoft Corporation. All rights reserved.
+.TAGS DSCConfiguration
+.LICENSEURI https://github.com/PowerShell/ComputerManagementDsc/blob/master/LICENSE
+.PROJECTURI https://github.com/PowerShell/ComputerManagementDsc
+.ICONURI
+.EXTERNALMODULEDEPENDENCIES
+.REQUIREDSCRIPTS
+.EXTERNALSCRIPTDEPENDENCIES
+.RELEASENOTES First version.
+.PRIVATEDATA 2016-Datacenter,2016-Datacenter-Server-Core
+#>
+
+#Requires -module ComputerManagementDsc
+
+<#
+    .DESCRIPTION
+        Example script that sets the application Windows Event Log
+        to a maximum size 4096MB, the logmode to 'Circular' and ensure that it is enabled.
+#>
+Configuration WindowsEventlog_SetWindowsEventlogSize_Config
+{
+    Import-DSCResource -ModuleName ComputerManagementDsc
+
+    Node localhost
+    {
+        WindowsEventLog ApplicationEventlogSize
+        {
+            LogName            = 'Application'
+            IsEnabled          = $true
+            LogMode            = 'Circular'
+            MaximumSizeInBytes = 4096KB
+        } # End of Windows Event Log Resource
+    } # End of Node
+} # End of Configuration